Refactor

+++ src/yamlService.ts
@@ -1,65 +1,89 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-const fs = require("fs");
-const YAML = require("yamljs");
-class YamlService {
-    ReadYaml(path) {
+import fs = require('fs');
+import YAML = require('yamljs');
+
+export default class YamlService {
+
+    public ReadYaml(path: string) {
+
         return YAML.load(path);
+
     }
-    AddEntryToYaml(entry, path) {
+
+    public AddEntryToYaml(entry: IQuestion, path: string) {
+
         const array = [];
         array[0] = entry;
         const entryAsYaml = YAML.stringify(array, undefined, 2);
+
         fs.appendFileSync(path, entryAsYaml);
+
     }
-    RemoveEntryFromYaml(entryIndex, path) {
+
+    public RemoveEntryFromYaml(entryIndex: number, path: string) {
+
         let yaml = this.ReadYaml(path);
+
         yaml.splice(entryIndex, 1);
         yaml = YAML.stringify(yaml, undefined, 2);
         fs.writeFileSync(path, yaml);
+
     }
-    EditEntryInYaml(entryIndex, path, question, answer) {
+
+    public EditEntryInYaml(entryIndex: number, path: string, question?: string, answer?: string) {
+
         const yamlEntry = this.ReadYaml(path)[entryIndex];
-        if (question) {
-            yamlEntry.question = question;
-        }
-        if (answer) {
-            yamlEntry.answer = answer;
-        }
+        if (question) { yamlEntry.question = question; }
+        if (answer) { yamlEntry.answer = answer; }
         this.RemoveEntryFromYaml(entryIndex, path);
         const array = [];
         array[0] = yamlEntry;
         const editedEntryAsYaml = YAML.stringify(array, undefined, 2);
         fs.appendFileSync(path, editedEntryAsYaml);
+
     }
-    CreateFile(path) {
+
+    public CreateFile(path: string): boolean {
+
         const isFileCreated = fs.existsSync(path);
-        if (!isFileCreated) {
-            fs.writeFileSync(path, '# YAML FILE\n');
-        }
+
+        if (!isFileCreated) { fs.writeFileSync(path, '# YAML FILE\n'); }
+
         return isFileCreated;
+
     }
-    DeleteFile(path) {
+
+    public DeleteFile(path: string): string {
+
         let result = `File ${path} deleted.`;
-        try {
-            fs.unlinkSync(path);
-        }
-        catch (_a) {
-            result = `File ${path} does not exist; no action taken.`;
-        }
+
+        try { fs.unlinkSync(path); } catch { result = `File ${path} does not exist; no action taken.`; }
+
         return result;
+
     }
-    CreateEntry(question) {
+
+    public CreateEntry(question: string): IQuestion {
+
         const DEFAULT_ANSWER = 'This question has not been answered yet.';
         const NULL_DATE = new Date(0);
         const CURRENT_DATE = new Date();
+
         return {
             question,
             answer: DEFAULT_ANSWER,
             dateOpened: CURRENT_DATE.toISOString(),
             dateClosed: NULL_DATE.toISOString(),
         };
+
     }
+
 }
-exports.default = YamlService;
-//# sourceMappingURL=app.js.map
+
+export interface IQuestion {
+
+    question: string;
+    answer: string;
+    dateOpened: string;
+    dateClosed: string;
+
+}
